README.md: Fix the Matrix address
There is a plumbed room at #feather:monero.social,
while #_oftc_#fether:matrix.org is a portal room. This is problematic
because it results in two Matrix rooms bridged to the same
IRC. #feather:monero.social has significantly more users, so switch to
that in the README to avoid new users joining the portal room.